Fix compiler warnings and deprecations for macOS 11+
- Adopted `NSMenuItemValidation` protocol in `GameController` to fix deprecated `validateMenuItem:` warning.
- Fixed `Arithmetic between different enumeration types` error in `Card.m` by casting enums to `NSInteger`.
- Removed unused `card` variable in `Game.m` to fix `Variable set but not used` warning.

+++ Card.m
@@ -65,7 +65,7 @@
 {
     // Suit ranges from 0 to 3; rank ranges from 1 (ACE) to 13 (KING).
     // This therefore returns a unique number between 0 and 51.
-    return (suit * KING + (rank - ACE));
+    return ((NSInteger)suit * (NSInteger)KING + ((NSInteger)rank - (NSInteger)ACE));
 }
 
 - (BOOL) isEqual: (id) other
